requirements.txt
langchain-google-genai
langchain-classic
langchain
langchain-chroma
langchainhub
dotenv
torch
transformers
pillow